backend /requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
python-dotenv==1.0.1
psycopg2==2.9.9
Flask-SQLAlchemy==3.1.1
scipy==1.14.1
numpy==2.1.1
